Rename modules in pep8 conform snake case
