+++ example/othello/logic.py
