+++ example/othello/nnet.py
